+++ client/src/pages/Legal/LegalNotice.jsx
@@ -1,0 +1,359 @@
+export default function LegalNotice() {
+  return (
+    <div className="pt-12 w-full px-8 sm:px-6 lg:px-14 2xl:px-80 bg-sky-50 -mb-4">
+      <section id="legal" className="mt-20">
+        <h5 className="text-sm text-zinc-600">
+          DERNIÈRE MISE À JOUR : 25-02-2025
+        </h5>
+        <h2 className="mt-2 mb-4 text-3xl font-bold border-b-2 font-kreon border-zinc-300">
+          MENTIONS LEGALES
+        </h2>
+
+        <p className="mb-4">
+          Conformément aux dispositions des{' '}
+          <a
+            href="https://www.legifrance.gouv.fr/loda/article_lc/LEGIARTI000044067469"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            Articles 6-III et 19 de la Loi n°2004-575 du 21 juin 2004 pour la
+            Confiance dans l’économie numérique, dite L.C.E.N.
+          </a>
+          , il est porté à la connaissance des Utilisateurs du site internet{' '}
+          <strong>https://luminescence33.com</strong> les présentes mentions
+          légales. La connexion et la navigation sur le site par l’Utilisateur
+          implique l’acceptation intégrale et sans réserve des présentes
+          mentions légales.
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 1 : L’éditeur
+        </h3>
+        <p className="mb-4">
+          L’édition du site LUMINESCENCE (https://luminescence33.com) est
+          assurée par la société <strong>LUMINESCENCE</strong> qui est une
+          micro-entreprise au capital de 0 euros, immatriculée Luminescence sous
+          le numéro 898 845 391 dont le siège social est situé au{' '}
+          <a
+            href="https://www.google.fr/maps/place/LUMINESCENCE/@44.638864,-1.0605077,19z/data=!3m1!4b1!4m6!3m5!1s0xd5499415bca91f7:0x1f53aae507216527!8m2!3d44.638863!4d-1.059864!16s%2Fg%2F11y43t5xvv?entry=ttu&g_ep=EgoyMDI0MDgyMS4wIKXMDSoASAFQAw%3D%3D"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            108 Place de la Claire, 2 ème étage (Ascenseur), 33470
+            Gujan-Mestras{' '}
+          </a>
+          .
+        </p>
+        <p className="mb-4">
+          <strong>Contact :</strong>
+          <br />
+          <a
+            href="mailto:contact@luminescence33.com"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            contact@luminescence33.com
+          </a>
+          <br />
+          <a
+            href="tel:+33 6 46 84 93 52"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            +33 6 46 84 93 52
+          </a>
+        </p>
+        <p className="mb-4">
+          <strong>Directeur de la publication : Pierre MAZELAYGUE -</strong>‍{' '}
+          <a
+            href="mailto:devteam@luminescence33.com"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            devteam@luminescence33.com
+          </a>
+          .
+        </p>
+        <p className="mb-4">
+          <strong>
+            Délégué à la protection des données :‍ Pierre MAZELAYGUE -
+          </strong>
+          ‍{' '}
+          <a
+            href="mailto:devteam@luminescence33.com"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            devteam@luminescence33.com
+          </a>
+          .
+        </p>
+        <p className="mb-4">
+          <strong>Webmaster : Pierre MAZELAYGUE -</strong>‍{' '}
+          <a
+            href="mailto:devteam@luminescence33.com"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            devteam@luminescence33.com
+          </a>
+          .
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 2 : L’hébergeur
+        </h3>
+        <p className="mb-4">
+          Le site LUMINESCENCE (https://luminescence33.com) est hébergé :
+        </p>
+        <p className="mb-4">
+          Sur les serveurs Hostinger UK Limited, UK
+          <br />
+          Hostinger UK Limited, UK private limited company
+          <br />
+          Nwms Center, 31 Southampton Row, Office 3.11, 3rd Floor
+          <br />
+          London, England, WC1B 5HJ.
+          <br />
+          <a
+            href="https://www.hostinger.fr/"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            ‍https://www.hostinger.fr/
+          </a>
+        </p>
+        <p className="mb-4">
+          Sur des centres de données de la société Hostinger UK Limited, UK
+          <br />
+          Hostinger UK Limited, UK private limited company
+          <br />
+          Nwms Center, 31 Southampton Row, Office 3.11, 3rd Floor
+          <br />
+          London, England, WC1B 5HJ.
+          <br />
+          <a
+            href="https://www.hostinger.fr/"
+            className="font-bold hover:underline text-sky-600"
+            target="_blank"
+            rel="noreferrer"
+          >
+            ‍https://www.hostinger.fr/
+          </a>
+          <br />
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 3 : Propriété intellectuelle et contrefaçon
+        </h3>
+        <p className="mb-4">
+          <strong>LUMINESCENCE</strong> est propriétaire des droits de propriété
+          intellectuelle et détient les droits d’usage sur tous les éléments
+          accessibles sur le site internet, notamment les textes, images,
+          graphismes, logos, vidéos, architecture, icônes et sons.
+        </p>
+        <p className="mb-4">
+          Toute reproduction, représentation, modification, publication,
+          adaptation de tout ou partie des éléments du site, quel que soit le
+          moyen ou le procédé utilisé, est interdite, sauf autorisation écrite
+          préalable de <strong>LUMINESCENCE</strong>.
+        </p>
+        <p className="mb-4">
+          Toute exploitation non autorisée du site ou de l’un quelconque des
+          éléments qu’il contient sera considérée comme constitutive d’une
+          contrefaçon et poursuivie conformément aux dispositions des articles{' '}
+          <a
+            href="https://www.legifrance.gouv.fr/codes/article_lc/LEGIARTI000032655082"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            L.335-2 et suivants du Code de Propriété Intellectuelle
+          </a>
+          .
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 4 : Limitation de responsabilité
+        </h3>
+        <p className="mb-4">
+          Le site est accessible par tout endroit, 7j/7, 24h/24 sauf cas de
+          force majeure, interruption programmée ou non et pouvant découler
+          d’une nécessité de maintenance.
+        </p>
+        <p className="mb-4">
+          <strong>LUMINESCENCE</strong> ne pourra être tenu pour responsable des
+          dommages directs et indirects causés au matériel de l’utilisateur,
+          lors de l’accès au site <strong>https://luminescence33.com</strong>.
+        </p>
+        <p className="mb-4">
+          <strong>LUMINESCENCE</strong> décline toute responsabilité quant à
+          l’utilisation qui pourrait être faite des informations et contenus
+          présents sur <strong>https://luminescence33.com</strong>.
+        </p>
+        <p className="mb-4">
+          <strong>LUMINESCENCE</strong> s’engage à sécuriser au mieux le site{' '}
+          <strong>https://luminescence33.com</strong>, cependant sa
+          responsabilité ne pourra être mise en cause si des données
+          indésirables sont importées et installées sur son site à son insu.
+        </p>
+        <p className="mb-4">
+          Des espaces interactifs (espace contact) sont à la disposition des
+          utilisateurs. <strong>LUMINESCENCE</strong> se réserve le droit de
+          supprimer, sans mise en demeure préalable, tout contenu déposé dans
+          cet espace qui contreviendrait à la législation applicable en France,
+          en particulier aux dispositions relatives à la protection des données,
+          cependant sa responsabilité ne pourra être mise en cause si des
+          données indésirables sont importées et installées sur son site à son
+          insu.
+        </p>
+        <p className="mb-4">
+          Le cas échéant <strong>LUMINESCENCE</strong> se réserve également la
+          possibilité de mettre en cause la responsabilité civile et/ou pénale
+          de l’utilisateur, notamment en cas de message à caractère raciste,
+          injurieux, diffamant, ou pornographique, quel que soit le support
+          utilisé (texte, photographie …). se réserve le droit de supprimer,
+          sans mise en demeure préalable, tout contenu déposé dans cet espace
+          qui contreviendrait à la législation applicable en France, en
+          particulier aux dispositions relatives à la protection des données.,
+          cependant sa responsabilité ne pourra être mise en cause si des
+          données indésirables sont importées et installées sur son site à son
+          insu.
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 5 : CNIL et gestion des données personnelles
+        </h3>
+        <p className="mb-4">
+          Conformément aux dispositions de{' '}
+          <a
+            href="https://www.legifrance.gouv.fr/loda/id/JORFTEXT000000886460"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            la loi 78-17 du 6 janvier 1978 modifiée
+          </a>
+          , l’utilisateur du site <strong>https://luminescence33.com</strong>{' '}
+          dispose d’un droit d’accès, de modification et de suppression des
+          informations collectées. Pour exercer ce droit, envoyez un message à
+          notre Délégué à la protection des données : Pierre MAZELAYGUE -‍{' '}
+          <a
+            href="mailto:devteam@luminescence33.com"
+            target="_blank"
+            rel="noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            devteam@luminescence33.com
+          </a>
+          .
+        </p>
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 6 : Liens hypertextes et cookies
+        </h3>
+        <p className="mb-4">
+          Le site <strong>https://luminescence33.com</strong> contient des liens
+          hypertextes vers d’autres sites et dégage toute responsabilité à
+          propos de ces liens externes ou des liens créés par d’autres sites
+          vers <strong>https://luminescence33.com</strong>.
+        </p>
+        <p className="mb-4">
+          Un &quot;cookie&quot; est un fichier de petite taille qui enregistre
+          des informations relatives à la navigation d’un utilisateur sur un
+          site. Les données ainsi obtenues permettent d&apos;obtenir des mesures
+          de fréquentation, améliorer l&apos;expérience utilisateur, par
+          exemple.
+        </p>
+        <p className="mb-4">
+          Vous avez la possibilité{' '}
+          <strong>d’accepter ou de refuser les cookies</strong> en modifiant les
+          paramètres de votre navigateur. Aucun cookie ne sera déposé sans votre
+          consentement. Un &quot;cookie&quot; est un fichier de petite taille
+          qui enregistre des informations relatives à la navigation d’un
+          utilisateur sur un site. Les données ainsi obtenues permettent
+          d&apos;obtenir des mesures de fréquentation, par exemple.
+        </p>
+        <p className="mb-4">
+          Les cookies sont enregistrés pour une durée maximale de 13 mois.
+        </p>
+        <p className="mb-4">
+          Pour plus d&apos;informations sur la façon dont nous faisons usage des
+          cookies, lisez notre{' '}
+          <a
+            href="/politique-confidentialite"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            politique de confidentialité
+          </a>
+          .
+        </p>
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 7 : Droit applicable et attribution de juridiction
+        </h3>
+        <p className="mb-4">
+          Tout litige en relation avec l’utilisation du site
+          <strong>https://luminescence33.com</strong> est soumis au droit
+          français. En dehors des cas où la loi ne le permet pas, il est fait
+          attribution exclusive de juridiction aux tribunaux compétents de{' '}
+          <strong>Nouvelle-Aquitaine</strong>.
+        </p>
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 8 : Demande de Réclamation
+        </h3>
+        <p className="mb-4">
+          Toute demande de réclamation concernant l&apos;agence ou/et le site
+          <strong>https://luminescence33.com</strong> doit être effectuée par
+          email. Veuillez envoyer votre réclamation à l&apos;adresse suivante :{' '}
+          <a
+            href="mailto:support@luminescence33.com"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            support@luminescence33.com
+          </a>
+          . Nous nous efforcerons de traiter votre demande dans les plus brefs.
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          ARTICLE 9 : Crédits Photos
+        </h3>
+        <p className="mb-4">
+          <strong>Photographe :</strong>{' '}
+          <a
+            href="https://www.facebook.com/profile.php?id=61553975133381"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            Karin LEGROS
+          </a>
+          .
+        </p>
+        <p className="mb-4">
+          Toutes demandes concernant les photos de l&apos;agence ou/et le site
+          <strong>https://luminescence33.com</strong> doit être effectuée par
+          email. Veuillez envoyer votre demande à l&apos;adresse suivante :{' '}
+          <a
+            href="mailto:support@luminescence33.com"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            support@luminescence33.com
+          </a>
+          . Nous nous efforcerons de traiter votre demande dans les plus brefs.
+        </p>
+      </section>
+    </div>
+  );
+}