+++ client/src/pages/Legal/PoliticyConfidentiality.jsx
@@ -1,0 +1,206 @@
+export default function PoliticyConfidentiality() {
+  return (
+    <div className="w-full px-4 pt-12 sm:px-6 lg:px-14 2xl:mx-80 bg-sky-50 -mb-4">
+      <section id="confidentialite" className="mt-20 ">
+        <h5 className="text-sm">DERNIÈRE MISE À JOUR : 25-02-2025</h5>
+        <h2 className="py-2 mb-4 text-3xl font-bold border-b-2 border-gray-200 font-kreon">
+          POLITIQUE DE CONFIDENTIALITÉ
+        </h2>
+        <p className="mb-4 font-bold">
+          <a
+            href="/"
+            rel="noopener noreferrer"
+            className="font-bold hover:underline text-sky-600"
+          >
+            https://luminescence33.com
+          </a>
+        </p>
+        <p className="mb-4 font-bold">LUMINESCENCE</p>
+        <p className="mb-4 ">
+          <strong>Type de site : </strong>Site Vitrine
+        </p>
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          Le but de cette politique de confidentialité
+        </h3>
+        <p className="mb-4">
+          Le but de cette politique de confidentialité est d&apos;informer les
+          utilisateurs de notre site des données personnelles que nous
+          recueillerons ainsi que les informations suivantes, le cas échéant :
+        </p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>Les données personnelles que nous recueillerons</li>
+          <li>L’utilisation des données recueillies</li>
+          <li>Qui a accès aux données recueillies</li>
+          <li>Les droits des utilisateurs du site</li>
+          <li>La politique de cookies du site</li>
+        </ul>
+        <p className="mb-4">
+          Cette politique de confidentialité fonctionne parallèlement aux
+          conditions générales d’utilisation de notre site.
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">Lois applicables</h3>
+        <p className="mb-4">
+          Conformément au Règlement général sur la protection des données
+          (RGPD), cette politique de confidentialité est conforme aux règlements
+          suivants.
+        </p>
+        <p className="mb-4">Les données à caractère personnel doivent être :</p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>
+            Traitées de manière licite, loyale et transparente au regard de la
+            personne concernée (licéité, loyauté, transparence) ;
+          </li>
+          <li>
+            Collectées pour des finalités déterminées, explicites et légitimes,
+            et ne pas être traitées ultérieurement d&apos;une manière
+            incompatible avec ces finalités; le traitement ultérieur à des fins
+            archivistiques dans l&apos;intérêt public, à des fins de recherche
+            scientifique ou historique ou à des fins statistiques n&apos;est pas
+            considéré, conformément à l&apos;article 89, paragraphe 1, comme
+            incompatible avec les finalités initiales (limitation des finalités)
+            ;
+          </li>
+          <li>
+            Adéquates, pertinentes et limitées à ce qui est nécessaire au regard
+            des finalités pour lesquelles elles sont traitées (minimisation des
+            données) ;
+          </li>
+          <li>
+            Exactes et, si nécessaire, tenues à jour; toutes les mesures
+            raisonnables doivent être prises pour que les données à caractère
+            personnel qui sont inexactes, eu égard aux finalités pour lesquelles
+            elles sont traitées, soient effacées ou rectifiées sans tarder
+            (exactitude) ;
+          </li>
+          <li>
+            Conservées sous une forme permettant l&apos;identification des
+            personnes concernées pendant une durée n&apos;excédant pas celle
+            nécessaire au regard des finalités pour lesquelles elles sont
+            traitées; les données à caractère personnel peuvent être conservées
+            pour des durées plus longues dans la mesure où elles seront traitées
+            exclusivement à des fins archivistiques dans l&apos;intérêt public,
+            à des fins de recherche scientifique ou historique ou à des fins
+            statistiques conformément à l&apos;article 89, paragraphe 1, pour
+            autant que soient mises en œuvre les mesures techniques et
+            organisationnelles appropriées requises par le règlement afin de
+            garantir les droits et libertés de la personne concernée (limitation
+            de la conservation) ;
+          </li>
+          <li>
+            Traitées de façon à garantir une sécurité appropriée des données à
+            caractère personnel, y compris la protection contre le traitement
+            non autorisé ou illicite et contre la perte, la destruction ou les
+            dégâts d&apos;origine accidentelle, à l&apos;aide de mesures
+            techniques ou organisationnelles appropriées (intégrité et
+            confidentialité).
+          </li>
+        </ul>
+        <p className="mb-4">
+          Le traitement n&apos;est licite que si, et dans la mesure où, au moins
+          une des conditions suivantes est remplie :
+        </p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>
+            La personne concernée a consenti au traitement de ses données à
+            caractère personnel pour une ou plusieurs finalités spécifiques ;
+          </li>
+          <li>
+            Le traitement est nécessaire à l&apos;exécution d&apos;un contrat
+            auquel la personne concernée est partie ou à l&apos;exécution de
+            mesures précontractuelles prises à la demande de celle-ci ;
+          </li>
+          <li>
+            Le traitement est nécessaire au respect d&apos;une obligation légale
+            à laquelle le responsable du traitement est soumis ;
+          </li>
+          <li>
+            Le traitement est nécessaire à la sauvegarde des intérêts vitaux de
+            la personne concernée ou d&apos;une autre personne physique ;
+          </li>
+          <li>
+            Le traitement est nécessaire à l&apos;exécution d&apos;une mission
+            d&apos;intérêt public ou relevant de l&apos;exercice de
+            l&apos;autorité publique dont est investi le responsable du
+            traitement ;
+          </li>
+          <li>
+            Le traitement est nécessaire aux fins des intérêts légitimes
+            poursuivis par le responsable du traitement ou par un tiers, à moins
+            que ne prévalent les intérêts ou les libertés et droits fondamentaux
+            de la personne concernée qui exigent une protection des données à
+            caractère personnel, notamment lorsque la personne concernée est un
+            enfant.
+          </li>
+        </ul>
+        <p className="mb-4">
+          Pour les résidents de l’État de Californie, cette politique de
+          confidentialité vise à se conformer à la California Consumer Privacy
+          Act (CCPA). S’il y a des incohérences entre ce document et la CCPA, la
+          législation de l’État s’appliquera. Si nous constatons des
+          incohérences, nous modifierons notre politique pour nous conformer à
+          la loi pertinente.
+        </p>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">Consentement</h3>
+        <p className="mb-4">
+          Les utilisateurs conviennent qu’en utilisant notre site, ils
+          consentent à :
+        </p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>
+            Les conditions énoncées dans la présente politique de
+            confidentialité et
+          </li>
+          <li>
+            La collecte, l’utilisation et la conservation des données énumérées
+            dans la présente politique.
+          </li>
+        </ul>
+
+        <h3 className="mt-8 mb-2 text-xl font-semibold">
+          Données personnelles que nous collectons
+        </h3>
+
+        <h4 className="mt-8 mb-2 text-lg font-semibold">
+          Données collectées automatiquement
+        </h4>
+        <p className="mb-4">
+          Lorsque vous visitez et utilisez notre site, nous pouvons
+          automatiquement recueillir et conserver les renseignements suivants :
+        </p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>Adresse IP</li>
+          <li>Détails matériels et logiciels</li>
+        </ul>
+
+        <h4 className="mt-8 mb-2 text-lg font-semibold">
+          Données recueillies de façon non automatique
+        </h4>
+        <p className="mb-4">
+          Nous pouvons également collecter les données suivantes lorsque vous
+          effectuez certaines fonctions sur notre site :
+        </p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>Prénom et nom / Entreprise</li>
+          <li>Email</li>
+          <li>Numéro de téléphone</li>
+        </ul>
+        <p className="mb-4">
+          Ces données peuvent être recueillies au moyen des méthodes suivantes :
+        </p>
+        <ul className="flex flex-col gap-2 mb-4 list-decimal list-inside">
+          <li>Formulaire de contact</li>
+          <li>Formulaire d&apos;entreprise</li>
+          <li>Formulaire de newsletter</li>
+        </ul>
+        <p className="mb-4">
+          Veuillez noter que nous ne collectons que les données qui nous aident
+          à atteindre l’objectif énoncé dans cette politique de confidentialité.
+          Nous ne recueillerons pas de données supplémentaires sans vous en
+          informer au préalable.
+        </p>
+      </section>
+    </div>
+  );
+}